// Flytec Front/Flytec/src/app/models/bula/bula.model.ts
export class Bula {
    idBula?: number;
    nomeProduto?: string;
    idCultura?: number;
    idClassificacaoToxicologica?: number;
    classe?: string;
    tipoDeFormulacao?: string;
    idAlvoBiologico?: number;
    doseProdutoComercial?: number;
    adjuvante?: string;
    idTipoDeServico?: number;
}